# Docs/refactor/artifacts/dashboard_openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SuiteTrialDetailV2": {
        "properties": {
          "architecture": {
            "additionalProperties": true,
            "title": "Architecture",
            "type": "object"
          },
          "architecture_summary": {
            "default": "",
            "title": "Architecture Summary",
            "type": "string"
          },
          "checkpoint_metadata": {
            "additionalProperties": true,
            "title": "Checkpoint Metadata",
            "type": "object"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "current_activity": {
            "additionalProperties": true,
            "title": "Current Activity",
            "type": "object"
          },
          "latest": {
            "additionalProperties": true,
            "title": "Latest",
            "type": "object"
          },
          "model_metadata": {
            "additionalProperties": true,
            "title": "Model Metadata",
            "type": "object"
          },
          "state": {
            "additionalProperties": true,
            "title": "State",
            "type": "object"
          },
          "trial": {
            "additionalProperties": true,
            "title": "Trial",
            "type": "object"
          },
          "trial_dir": {
            "title": "Trial Dir",
            "type": "string"
          },
          "trial_id": {
            "title": "Trial Id",
            "type": "string"
          }
        },
        "required": [
          "trial_id",
          "trial_dir"
        ],
        "title": "SuiteTrialDetailV2",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Hexo-RL Dashboard",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/suite/best-checkpoints": {
      "get": {
        "operationId": "best_checkpoints_api_suite_best_checkpoints_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Best Checkpoints Api Suite Best Checkpoints Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Best Checkpoints"
      }
    },
...
    "/api/suite/trials/{trial_id}": {
      "get": {
        "operationId": "trial_detail_api_suite_trials__trial_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trial_id",
            "required": true,
            "schema": {
              "title": "Trial Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuiteTrialDetailV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trial Detail"
      }
    },
    "/api/suite/trials/{trial_id}/events": {
      "get": {
        "operationId": "trial_events_api_suite_trials__trial_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "trial_id",
            "required": true,
            "schema": {
              "title": "Trial Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Trial Events Api Suite Trials  Trial Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trial Events"
      }
    },
    "/api/suite/trials/{trial_id}/loss-curve": {
      "get": {
        "operationId": "trial_loss_curve_api_suite_trials__trial_id__loss_curve_get",
        "parameters": [
          {
            "in": "path",
            "name": "trial_id",
            "required": true,
            "schema": {
              "title": "Trial Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Trial Loss Curve Api Suite Trials  Trial Id  Loss Curve Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trial Loss Curve"
      }
    },
    "/api/suite/trials/{trial_id}/runtime-sweep": {
      "get": {
        "operationId": "trial_runtime_sweep_api_suite_trials__trial_id__runtime_sweep_get",
        "parameters": [
          {
            "in": "path",
            "name": "trial_id",
            "required": true,
            "schema": {
              "title": "Trial Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Trial Runtime Sweep Api Suite Trials  Trial Id  Runtime Sweep Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trial Runtime Sweep"
      }
    },
    "/api/suite/trials/{trial_id}/scores": {
      "get": {
        "operationId": "trial_scores_api_suite_trials__trial_id__scores_get",
        "parameters": [
          {
            "in": "path",
            "name": "trial_id",
            "required": true,
            "schema": {
              "title": "Trial Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Trial Scores Api Suite Trials  Trial Id  Scores Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trial Scores"
      }
    }
  }
}